fixed old typings
