+++ src/lang/pt_br.d.ts
@@ -1,5 +1,5 @@
-import { Lang } from 'suneditor';
+import { Lang } from './Lang';
 
-const ptBr: Lang;
+declare const ptBr: Lang;
 
 export default ptBr;
