+++ src/options.d.ts
@@ -1,18 +1,8 @@
 import { DisplayProperty, PositionProperty } from 'csstype';
+import { Lang } from './lang/Lang';
+import { Plugin } from './plugins/Plugin';
 
-declare module 'suneditor' {
-  export function create(
-    idOrElement: string | Element,
-    options: SunEditorOptions,
-    _init_options?: SunEditorOptions
-  ): any;
-
-  export function init(init_options: SunEditorOptions): any;
-
-  /**
-   * SunEditor Options
-   */
-  export interface SunEditorOptions {
+export interface SunEditorOptions {
     plugins: Plugin[];
     /**
      * Add tags to the default tags whitelist of editor. default: '' {String}
@@ -313,204 +303,56 @@ declare module 'suneditor' {
      */
     icons?: Record<string, string>;
   }
-
   export type FormatTagName =
-    'p'
-    | 'div'
-    | 'blockquote'
-    | 'pre'
-    | 'h1'
-    | 'h2'
-    | 'h3'
-    | 'h4'
-    | 'h5'
-    | 'h6'
-    | { tag: string; name?: string; command: 'replace' | 'range' | 'free'; class: string };
-  export type TemplatesItem = { name: string; html: string };
-  export type LineHeightsItem = { text: string; value: number };
-  export type ParagraphStyles = (string | { name: string; class: string; _class?: string })[];
-  export type TextStyles = (string | { name: string; style: string; tag: string; _class?: string })[];
-  export type ButtonListDefaults =
-    | '/' // Line Break
-    | 'bold'
-    | 'underline'
-    | 'italic'
-    | 'strike'
-    | 'subscript'
-    | 'superscript'
-    | 'removeFormat'
-    | 'indent'
-    | 'outdent'
-    | 'fullScreen'
-    | 'showBlocks'
-    | 'codeView'
-    | 'undo'
-    | 'redo'
-    | 'preview'
-    | 'print'
-    | 'save'
-    | 'font'
-    | 'formatBlock'
-    | 'fontSize'
-    | 'fontColor'
-    | 'hiliteColor'
-    | 'align'
-    | 'list'
-    | 'horizontalRule'
-    | 'table'
-    | 'lineHeight'
-    | 'template'
-    | 'paragraphStyle'
-    | 'textStyle'
-    | 'link'
-    | 'image'
-    | 'video'
-    | 'math';
-  export type ButtonListItem = ButtonListDefaults | ButtonListDefaults[];
-
-  export interface Plugin {
-    name: string;
-    display: string;
-    add: (core: any, targetElement?: any) => void;
-    active: (element: any) => boolean;
-  }
-
-  export interface CommandPlugin extends Plugin {
-    action: () => void;
-  }
-
-  export interface SubmenuPlugin extends Plugin {
-    on: () => void;
-  }
-
-  export interface DialogPlugin extends Plugin {
-    open: () => void;
-    on: () => void;
-    init: () => void;
-  }
-
-  export interface Lang {
-    toolbar: {
-      default: string,
-      save: string,
-      font: string,
-      formats: string,
-      fontSize: string,
-      bold: string,
-      underline: string,
-      italic: string,
-      strike: string,
-      subscript: string,
-      superscript: string,
-      removeFormat: string,
-      fontColor: string,
-      hiliteColor: string,
-      indent: string,
-      outdent: string,
-      align: string,
-      alignLeft: string,
-      alignRight: string,
-      alignCenter: string,
-      alignJustify: string,
-      list: string,
-      orderList: string,
-      unorderList: string,
-      horizontalRule: string,
-      hr_solid: string,
-      hr_dotted: string,
-      hr_dashed: string,
-      table: string,
-      link: string,
-      math: string,
-      image: string,
-      video: string,
-      fullScreen: string,
-      showBlocks: string,
-      codeView: string,
-      undo: string,
-      redo: string,
-      preview: string,
-      print: string,
-      tag_p: string,
-      tag_div: string,
-      tag_h: string,
-      tag_blockquote: string,
-      tag_pre: string,
-      template: string,
-      lineHeight: string,
-      paragraphStyle: string,
-      textStyle: string
-    };
-    dialogBox: {
-      linkBox: {
-        title: string,
-        url: string,
-        text: string,
-        newWindowCheck: string
-      },
-      mathBox: {
-        title: string,
-        inputLabel: string,
-        fontSizeLabel: string,
-        previewLabel: string
-      },
-      imageBox: {
-        title: string,
-        file: string,
-        url: string,
-        altText: string
-      },
-      videoBox: {
-        title: string,
-        url: string
-      },
-      caption: string,
-      close: string,
-      submitButton: string,
-      revertButton: string,
-      proportion: string,
-      basic: string,
-      left: string,
-      right: string,
-      center: string,
-      width: string,
-      height: string,
-      size: string,
-      ratio: string
-    };
-    controller: {
-      edit: string,
-      unlink: string,
-      remove: string,
-      insertRowAbove: string,
-      insertRowBelow: string,
-      deleteRow: string,
-      insertColumnBefore: string,
-      insertColumnAfter: string,
-      deleteColumn: string,
-      resize100: string,
-      resize75: string,
-      resize50: string,
-      resize25: string,
-      autoSize: string,
-      mirrorHorizontal: string,
-      mirrorVertical: string,
-      rotateLeft: string,
-      rotateRight: string,
-      maxSize: string,
-      minSize: string,
-      tableHeader: string,
-      mergeCells: string,
-      splitCells: string,
-      HorizontalSplit: string,
-      VerticalSplit: string
-    };
-    menu: {
-      spaced: string,
-      bordered: string,
-      neon: string,
-      translucent: string,
-      shadow: string
-    };
-  }
-}
+  'p'
+  | 'div'
+  | 'blockquote'
+  | 'pre'
+  | 'h1'
+  | 'h2'
+  | 'h3'
+  | 'h4'
+  | 'h5'
+  | 'h6'
+  | { tag: string; name?: string; command: 'replace' | 'range' | 'free'; class: string };
+export type TemplatesItem = { name: string; html: string };
+export type LineHeightsItem = { text: string; value: number };
+export type ParagraphStyles = (string | { name: string; class: string; _class?: string })[];
+export type TextStyles = (string | { name: string; style: string; tag: string; _class?: string })[];
+export type ButtonListDefaults =
+  | '/' // Line Break
+  | 'bold'
+  | 'underline'
+  | 'italic'
+  | 'strike'
+  | 'subscript'
+  | 'superscript'
+  | 'removeFormat'
+  | 'indent'
+  | 'outdent'
+  | 'fullScreen'
+  | 'showBlocks'
+  | 'codeView'
+  | 'undo'
+  | 'redo'
+  | 'preview'
+  | 'print'
+  | 'save'
+  | 'font'
+  | 'formatBlock'
+  | 'fontSize'
+  | 'fontColor'
+  | 'hiliteColor'
+  | 'align'
+  | 'list'
+  | 'horizontalRule'
+  | 'table'
+  | 'lineHeight'
+  | 'template'
+  | 'paragraphStyle'
+  | 'textStyle'
+  | 'link'
+  | 'image'
+  | 'video'
+  | 'math';
+export type ButtonListItem = ButtonListDefaults | ButtonListDefaults[];
