+++ src/plugins/Plugin.d.ts
@@ -1,0 +1,6 @@
+export interface Plugin {
+    name: string;
+    display: string;
+    add: (core: any, targetElement?: any) => void;
+    active: (element: any) => boolean;
+}
