+++ src/plugins/command/CommandPlugin.d.ts
@@ -1,0 +1,5 @@
+import { Plugin } from '../Plugin';
+
+export interface CommandPlugin extends Plugin {
+    action: () => void;
+}
