+++ src/plugins/dialog/DialogPlugin.d.ts
@@ -1,0 +1,7 @@
+import { Plugin } from '../Plugin';
+
+export interface DialogPlugin extends Plugin {
+    open: () => void;
+    on: () => void;
+    init: () => void;
+}
