+++ src/plugins/submenu/SubmenuPlugin.d.ts
@@ -1,0 +1,5 @@
+import { Plugin } from '../Plugin';
+
+export interface SubmenuPlugin extends Plugin {
+    on: () => void;
+}
