+++ src/suneditor.d.ts
@@ -1,5 +1,8 @@
+import { SunEditorOptions } from './options.d';
+
 declare namespace _default {
-    export function init(init_options: any): any;
-    export function create(idOrElement: String | Element, options: any, _init_options: any): any;
+    export function init(init_options: SunEditorOptions): any;
+    export function create(idOrElement: String | Element, options: SunEditorOptions, _init_options?: SunEditorOptions): any;
 }
+
 export default _default;
